tendermint-rpc: Rename 0.33 client files
